+++ BitFaster.Caching.Benchmarks/Lfu/SketchIncrement.cs
@@ -16,18 +16,6 @@ namespace BitFaster.Caching.Benchmarks.Lfu
         private static CmSketch<int, DisableHardwareIntrinsics> std = new CmSketch<int, DisableHardwareIntrinsics>(sketchSize, EqualityComparer<int>.Default);
         private static CmSketch<int, DetectIsa> avx = new CmSketch<int, DetectIsa>(sketchSize, EqualityComparer<int>.Default);
 
-        private static CmSketchBlock<int, DisableHardwareIntrinsics> block = new CmSketchBlock<int, DisableHardwareIntrinsics>(sketchSize, EqualityComparer<int>.Default);
-        private static CmSketchBlock<int, DetectIsa> blockAvx = new CmSketchBlock<int, DetectIsa>(sketchSize, EqualityComparer<int>.Default);
-
-        private static CmSketchBlockV2<int, DisableHardwareIntrinsics> block2 = new CmSketchBlockV2<int, DisableHardwareIntrinsics>(sketchSize, EqualityComparer<int>.Default);
-        private static CmSketchBlockV2<int, DetectIsa> block2Avx = new CmSketchBlockV2<int, DetectIsa>(sketchSize, EqualityComparer<int>.Default);
-
-        private static CmSketchBlockSegmentRemoved<int, DisableHardwareIntrinsics> blockSeg = new CmSketchBlockSegmentRemoved<int, DisableHardwareIntrinsics>(sketchSize, EqualityComparer<int>.Default);
-        private static CmSketchBlockSegmentRemoved<int, DetectIsa> blockSegAvx = new CmSketchBlockSegmentRemoved<int, DetectIsa>(sketchSize, EqualityComparer<int>.Default);
-
-        private static CmSketchBlockSegmentRemoved<int, DisableHardwareIntrinsics> blockSegRem = new CmSketchBlockSegmentRemoved<int, DisableHardwareIntrinsics>(sketchSize, EqualityComparer<int>.Default);
-        private static CmSketchBlockSegmentRemoved<int, DetectIsa> blockSegRemAvx = new CmSketchBlockSegmentRemoved<int, DetectIsa>(sketchSize, EqualityComparer<int>.Default);
-
         [Benchmark(Baseline = true, OperationsPerInvoke = iterations)]
         public void Inc()
         {
@@ -43,78 +31,6 @@ namespace BitFaster.Caching.Benchmarks.Lfu
             for (int i = 0; i < iterations; i++)
             {
                 avx.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlock()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                block.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlockAvx()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                blockAvx.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlock2()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                block2.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlock2Avx()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                block2Avx.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlockSeg()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                blockSeg.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlockSegAvx()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                blockSegAvx.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlockSegRem()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                blockSegRem.Increment(i);
-            }
-        }
-
-        [Benchmark(OperationsPerInvoke = iterations)]
-        public void IncBlockSegRemAvx()
-        {
-            for (int i = 0; i < iterations; i++)
-            {
-                blockSegRemAvx.Increment(i);
             }
         }
     }
